+++ bepensa-models/DataModels/RequestByIdUsuario.cs
@@ -6,6 +6,6 @@ namespace bepensa_models.DataModels
 {
     [Display(Name = "IdUsuario")]
     [Required(ErrorMessage = "El campo {0} es obligatorio")]
-    public long IdUsuario { get; set; }
+    public int IdUsuario { get; set; }
 }
 }
